Update train_alcnet.py
Relocate all the parameter files and training log files to the /params folder.
Improve logging message, now includes all the hyperparameter information for each run.
Time stamp for each message.
Date prefix for all the saved file.
If using in Colab, the time recorded is the system time of the remote host.